Add a default wp-cli settings file
Because hm-base uses a non-standard WordPress layout with the WordPress files in /wordpress/ and the WP_CONTENT_DIR set to /content/ wp-cli doesn't work out of the box when run from the site root. By defining the path parameter in this wp-cli file we can override that default behaviour and wp-cli will work.