Rename `plugins-mu` back to WordPress default.
A Slack discussion raised the question of the origin of this naming convention,
and we couldn't find a good reason to keep it. Reasons to change are that it is
a customisation that doesn't seem to benefit the majority of our current
projects, and for terminal users, it makes autocomplete a bit more frustrating.